+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Add hint to work around the "too many open files" error (issue #52).
 - Add warning for mismatch of dependency name and name in package
 - Add dependency information to `sources` command
+- Add `--fetch/-f` argument to `bender update` to force re-fetch of git dependencies from their remotes
+- Add global `--local` argument to disable remote accesses of git commands, e.g. for air-gapped computers
 
 ### Changed
 - Reduce the number of open files in large repositories by changing the method to get the Git commit hash from a tag (from individual calls to `git rev-parse --verify HASH^{commit}` to `git show-ref --dereference`).